chore: move `Claims` struct to dedicated `jwt` module in `models` in `auth` app
