+++ apps/auth/src/models/mod.rs
@@ -1,4 +1,5 @@
 pub mod auth;
+pub mod jwt;
 mod user;
 
 pub use user::*;
